requirements.txt
environs==9.3.3
python-telegram-bot==11.1.0
redis==3.5.3
vk-api==11.9.6